Change swappy for satty

+++ .config/hypr/scripts/screenshot.sh
@@ -4,12 +4,15 @@
 # | Simple Screenshot |
 #  -------------------
 temp_file=$(mktemp)
+scr_dir=~/Screenshots
 pkill -x slurp || grim -g "$(slurp)" - > $temp_file 
-file_name=$(date +'screenshot-%Y%m%d-%H%M.png')
-swappy -f $temp_file -o "$file_name"
-rm $temp_file
-rm $file_name
 
-if [[ -e ~/Imagenes/${file_name} ]]; then
-    notify-send -a "Swappy" "Screenshot" "Saved in $HOME/Imagenes/${file_name}" -i ~/Imagenes/${file_name} -r 91190 -t 8000
+if [[ ! -d $scr_dir ]]; then mkdir -p $scr_dir; fi
+
+if [[ -s "$temp_file" ]]; then
+    file_name=$(date +'screenshot-%Y%m%d-%H%M.png')
+    satty -f $temp_file -o "$scr_dir/$file_name"
+    rm $temp_file
+else
+    exit 0
 fi
